requirements-dev.txt
mock
pytest>=3.6
pytest-cov
tox
behave
pexpect
coverage
codecov
black
wheel
autoflake
testcontainers-iris
